requirements.txt
alembic==1.16.2
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown==3.4.4
MarkupSafe==3.0.2
Pillow==10.0.1
pyotp==2.9.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.32.4
soupsieve==2.7
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==2.3.7
WTForms==3.1.0
PyMySQL==1.1.1
selenium==4.15.2
webdriver-manager==4.0.1
boto3==1.35.0
